+++ src/main/java/roomescape/admin/AdminController.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.admin;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -6,10 +6,15 @@ import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
 @RequestMapping("/admin")
-public class RoomescapeController {
+public class AdminController {
 
-    @GetMapping()
+    @GetMapping
     public String getAdminPage() {
         return "admin/index";
+    }
+
+    @GetMapping("/reservation")
+    public String getReservationPage() {
+        return "admin/reservation-legacy";
     }
 }
