+++ src/main/java/roomescape/reservation/ReservationController.java
@@ -5,33 +5,28 @@ import java.util.List;
 import java.util.concurrent.atomic.AtomicLong;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 @Controller
+@RequestMapping("/reservations")
 public class ReservationController {
 
     private final AtomicLong index = new AtomicLong(1L);
-    private List<Reservation> reservations = new ArrayList<>();
+    private final List<Reservation> reservations = new ArrayList<>();
 
-    @GetMapping("/admin/reservation")
-    public String getReservationPage(Model model) {
-        return "admin/reservation-legacy";
-    }
-
-    @GetMapping("/reservations")
+    @GetMapping
     @ResponseBody
     public List<Reservation> getReservations() {
         return reservations;
     }
 
-
-    @PostMapping("/reservations")
+    @PostMapping
     @ResponseBody
     public ResponseEntity<Reservation> addReservation(@RequestBody ReservationDto reservationDto) {
         Reservation reservation = new Reservation(index.getAndIncrement(), reservationDto.name(), reservationDto.date(),
@@ -40,7 +35,7 @@ public class ReservationController {
         return ResponseEntity.ok(reservation);
     }
 
-    @DeleteMapping("/reservations/{id}")
+    @DeleteMapping("/{id}")
     @ResponseBody
     public ResponseEntity<Void> getReservations(@PathVariable("id") long id) {
         boolean isDeleted = reservations.removeIf((reservation) -> reservation.getId() == id);
